webcam_qr/demo/requirements.txt
gradio_webcam_qr